fix(core.deployment): dpa.properties is written even if package is already in persistence (#4690)
* fix(core.deployment): dpa.properties is written even if package is already in persistence


* fix(core.deployment.agent): use '-' as version separator, always try to update dpa.properties


* fix: Delete downloaded dp only if exist, otherwise exception is thrown


* fix: Use underscore for dps and dash for scripts


* fix(tests): fixed tests


* revert: stored name is constructed from DP info from deploymentAgent and not from metrics


* fix: Attempt to fix test


---------

+++ kura/org.eclipse.kura.deployment.agent/src/main/java/org/eclipse/kura/deployment/agent/impl/DeploymentAgent.java
@@ -380,13 +380,15 @@ public class DeploymentAgent implements DeploymentAgentService, ConfigurableComp
             if (!dpFile.getCanonicalPath().equals(dpPersistentFile.getCanonicalPath())) {
                 logger.debug("dpFile.getCanonicalPath(): {}", dpFile.getCanonicalPath());
                 logger.debug("dpPersistentFile.getCanonicalPath(): {}", dpPersistentFile.getCanonicalPath());
-                FileUtils.copyFile(dpFile, dpPersistentFile);
-                addPackageToConfFile(dp.getName(), "file:" + dpPersistentFilePath);
+                Files.deleteIfExists(dpPersistentFile.toPath());
+                FileUtils.moveFile(dpFile, dpPersistentFile);
             }
+
+            addPackageToConfFile(dp.getName(), "file:" + dpPersistentFilePath);
         } finally {
             // The file from which we have installed the deployment package will be deleted
             // unless it's a persistent deployment package file.
-            if (dpPersistentFile != null && dpPersistentFile.exists()
+            if (dpPersistentFile != null && dpPersistentFile.exists() && dpFile.exists()
                     && !dpFile.getCanonicalPath().equals(dpPersistentFile.getCanonicalPath())) {
                 Files.delete(dpFile.toPath());
                 logger.debug("Deleted file: {}", dpFile.getName());
